Fix bug causing mismatch between ``Order.user`` and ``Basket.owner``
During order placement, if the order ownership calculator assigns the
order to a user other than the basket owner, the basket would not be
updated to also have the new owner. Now, after creating the order model,
the owner of the basket associated with the order is updated to match
the order's owner.