Fix bug with multi-step payment methods when retrying a payment decline
The checkout serializer was reassigning basket ownership to the order owner as
soon as the order was placed. This broke retrying payments since the API would
then return the request user a new basket, not the basket belonging to the order
they want to retry.

This fixes the issue by waiting to reassign basket ownership until after the
order payment is complete. Order ownership is still transferred upon order
placement.